new impl follow array/map/struct
