+++ fe/fe-common/src/main/java/org/apache/doris/catalog/Type.java
@@ -381,13 +381,13 @@ public abstract class Type {
     // 3. don't support group by
     // 4. don't support index
     public boolean isOnlyMetricType() {
-        return isObjectStored() || isComplexType();
+        return isObjectStored() || isComplexType() || isJsonbType();
     }
 
     public static final String OnlyMetricTypeErrorMsg =
-            "Doris hll, bitmap, array, map, struct column must use with specific function, and don't"
+            "Doris hll, bitmap, array, map, struct, jsonb column must use with specific function, and don't"
                     + " support filter or group by. please run 'help hll' or 'help bitmap' or 'help array'"
-                    + " or 'help map' or 'help struct' in your mysql client.";
+                    + " or 'help map' or 'help struct' or 'help jsonb' in your mysql client.";
 
     public boolean isHllType() {
         return isScalarType(PrimitiveType.HLL);
@@ -525,21 +525,6 @@ public abstract class Type {
      */
     public boolean supportsTablePartitioning() {
         return false;
-    }
-
-    /**
-     * Indicates whether this type can be used as key of table.
-     */
-    public boolean supportsTableKey() {
-        return supportsComparison();
-    }
-
-    /**
-     * Indicates whether this type can be used as key of table.
-     */
-    public boolean supportsComparison() {
-        return !(isHllType() || isBitmapType() || isJsonbType()
-                    || isArrayType() || isMapType() || isStructType());
     }
 
     public PrimitiveType getPrimitiveType() {
