+++ fe/fe-core/src/main/java/org/apache/doris/analysis/MVColumnItem.java
@@ -117,11 +117,7 @@ public class MVColumnItem {
         this.type = type;
     }
 
-    public void setIsKey(boolean key) throws AnalysisException {
-        if (!type.supportsTableKey()) {
-            throw new AnalysisException("materialized view column " + name
-                        + " of type " + type + " can not be key");
-        }
+    public void setIsKey(boolean key) {
         isKey = key;
     }
